refactor: move window to its own module
