+++ Cargo.toml
@@ -14,4 +14,3 @@ x11 = { version = "2", features = ["xlib"] }
 pop-launcher = "=1.0.3"
 serde_json = "1.0.70"
 once_cell = "1.8.0"
-wnck-sys = "0.1.0"
