--- examples/gtklauncher/window/.#window.ui
@@ -1,1 +1,0 @@
-wash@pop-os.22535:1638129315
