+++ examples/gtklauncher/window/window.ui
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <interface>
-  <template class="TodoWindow" parent="GtkApplicationWindow">
-    <property name="width-request">360</property>
-    <property name="title" translatable="yes">To-Do</property>
+  <template class="LauncherWindow" parent="GtkApplicationWindow">
+    <property name="width-request">600</property>
+    <property name="title">Gtk Pop Launcher</property>
+    <property name="decorated">false</property>
     <child>
       <object class="GtkBox">
         <property name="orientation">vertical</property>
@@ -10,14 +11,15 @@
         <property name="margin-bottom">12</property>
         <property name="margin-start">12</property>
         <property name="margin-end">12</property>
-        <property name="spacing">6</property>
         <child>
-          <object class="GtkEntry" id="entry"/>
+          <object class="GtkEntry" id="entry">
+            <property name="margin-bottom">12</property>
+          </object>
         </child>
         <child>
           <object class="GtkScrolledWindow">
             <property name="hscrollbar-policy">never</property>
-            <property name="min-content-height">360</property>
+            <property name="min-content-height">500</property>
             <property name="vexpand">true</property>
             <child>
               <object class="GtkListView" id="list_view"/>
